MoveBtn class binding, instantiated pause button classes + added joystick fixes
this was hell pls it took me way too long to do that garbage

as for joystick fixes, the problem lies in making the event 'click' instead of mousedown or touchstart. the latter in particular causes problems because it registers the touchstart as the second touchstart as well; as soon as i pause, it immediately unpauses.
but, if i use click, i cant press pause while using the joystick. it doesnt register as a click

so easy solution: when the joystick is used,  add a touchstart event listener, and remove it upon joystick stop.
for redunduncy, i also disabled it two other times in the pauseFunction itself, and the joystickPauseFunction which communicates the joystick and the pauseBtn and also joystickPauseFunction stops the joystick;
solution might be super trashy but i mean it works i think

minor changes with css, i wanna remove <style> all together for everything thats made with javascript

i prolly shouldnt have jammed so much in one commit but whoops